Update Customer/Create Products
Creating a new product now checks to see if its brand and category is already in the database before inserting it.

A smilar process has been created for inserting new addresses as well as checking the address junction table to delete the old address ID before updating the table.

The addresses table also needed nullable added to a few  fields.